Create new Publisher class for pubsub messages.
